+++ serde_arrow/src/test_with_arrow/impls/arrow_dictionary.rs
@@ -182,7 +182,21 @@ mod construction {
             values: Box::new(to_array(DataType::Utf8, true, [None::<&str>, None])),
         });
 
-        assert!(ArrayDeserializer::new(String::from("$"), None, array.as_view()).is_err());
+        let deserializer =
+            ArrayDeserializer::new(String::from("$"), None, array.as_view()).unwrap();
+
+        assert_eq!(
+            Option::<String>::deserialize(deserializer.at(0)).unwrap(),
+            None
+        );
+        assert_eq!(
+            Option::<String>::deserialize(deserializer.at(1)).unwrap(),
+            None
+        );
+        assert_eq!(
+            Option::<String>::deserialize(deserializer.at(2)).unwrap(),
+            None
+        );
     }
 
     #[test]
@@ -196,34 +210,44 @@ mod construction {
             )),
         });
 
-        assert!(ArrayDeserializer::new(String::from("$"), None, array.as_view()).is_err());
+        let deserializer =
+            ArrayDeserializer::new(String::from("$"), None, array.as_view()).unwrap();
+
+        assert_eq!(
+            Option::<String>::deserialize(deserializer.at(0)).unwrap(),
+            None
+        );
+        assert_eq!(
+            Option::<String>::deserialize(deserializer.at(1)).unwrap(),
+            None
+        );
+        assert_eq!(
+            Option::<String>::deserialize(deserializer.at(2)).unwrap(),
+            Some(String::from("foo")),
+        );
     }
 
     #[test]
-    fn some_null_values_v2() {
+    fn some_null_keys() {
         let array = Array::Dictionary(DictionaryArray {
-            keys: Box::new(to_array(DataType::Int8, false, [1, 1, 0])),
-            values: Box::new(to_array(
-                DataType::Utf8,
-                true,
-                [
-                    Some("1"),
-                    Some("1"),
-                    Some("1"),
-                    Some("1"),
-                    Some("1"),
-                    Some("1"),
-                    Some("1"),
-                    Some("1"),
-                    Some("1"),
-                    Some("1"),
-                    Some("1"),
-                    None,
-                    Some("bar"),
-                ],
-            )),
+            keys: Box::new(to_array(DataType::Int8, true, [Some(1), None, Some(0)])),
+            values: Box::new(to_array(DataType::Utf8, false, ["foo", "bar"])),
         });
 
-        assert!(ArrayDeserializer::new(String::from("$"), None, array.as_view()).is_err());
+        let deserializer =
+            ArrayDeserializer::new(String::from("$"), None, array.as_view()).unwrap();
+
+        assert_eq!(
+            Option::<String>::deserialize(deserializer.at(0)).unwrap(),
+            Some(String::from("bar")),
+        );
+        assert_eq!(
+            Option::<String>::deserialize(deserializer.at(1)).unwrap(),
+            None
+        );
+        assert_eq!(
+            Option::<String>::deserialize(deserializer.at(2)).unwrap(),
+            Some(String::from("foo")),
+        );
     }
 }