+++ Changes.md
@@ -2,6 +2,14 @@
 
 ## `development`
 
+- Support deserializing dictionary encoded arrays with null values
+
+## Thanks
+
+- [@bennetthardwick](https://github.com/bennetthardwick) added support to
+  deserialize dictionaries with null values
+  ([#309](https://github.com/chmp/serde_arrow/pull/309))
+
 ## 0.14.1
 
 - Support `_` separators when serializing strings to "Float16, Float32,